requirements.txt
requests
PyYAML
ttkbootstrap
pytest
